requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14
cryptography>=41.0
python-dotenv>=1.0
psycopg2-binary>=2.9
pytest>=7.4
pytest-django>=4.7
